@@ -81,7 +81,7 @@ class DiffMatchPatch {
    * Returns a List of [Diff] objects.
    */
   List<d.Diff> diff(String text1, String text2,
-                    [bool checklines = true, DateTime deadline]) {
+                    [bool checklines = true, DateTime? deadline]) {
     return d.diff(text1, text2, checklines: checklines, deadline: deadline,
         timeout: diffTimeout);
   }
@@ -149,7 +149,7 @@ class DiffMatchPatch {
    *
    * Returns a List of Patch objects.
    */
-  List<p.Patch> patch(Object a, [Object opt_b, Object opt_c]) {
+  List<p.Patch> patch(Object a, [Object? opt_b, Object? opt_c]) {
     return p.patchMake(a, b: opt_b, c: opt_c, diffTimeout: diffTimeout,
         diffEditCost: diffEditCost, deleteThreshold: patchDeleteThreshold,
         margin: patchMargin);
